% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network}
\alias{build_network}
\title{Build a network
Builds a network from time series data using vector autoregression}
\usage{
build_network(X, exogen)
}
\arguments{
\item{X}{matrix with columns for species and each row is a time-ordered observation}

\item{exogen}{matrix of exogenous variables with the same number of rows as \code{X}}
}
\value{
A list with two elements. The first element is the \code{graph}, an iGraph object. The second element is a length-3 list containing the adjacency matrix (\code{adj_mat}), the p-values for that matrix (\code{adj_mat_p}), and the vector autoregression model from \code{vars::VAR} (\code{var_mod}).
}
\description{
Build a network
Builds a network from time series data using vector autoregression
}

