% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robnet.R
\name{print.robnet}
\alias{print.robnet}
\title{Print the network
Print the network, doesn't show attributes}
\usage{
\method{print}{robnet}(x)
}
\arguments{
\item{x}{output from \code{\link{build_network}}}
}
\description{
Print the network
Print the network, doesn't show attributes
}

